Moved participation.go to aws/participationManager.go
