+++ internal/participation_aggregation/aggregate.go
@@ -2,7 +2,7 @@ package participation_aggregation
 
 import "context"
 
-type ParticipationManager interface {
+type CrowdactionParticipationManager interface {
 	ChangeCrowdactionParticipantCountBy(ctx context.Context, crowdactionID string, count int) error
 }
 
@@ -11,13 +11,13 @@ type Service interface {
 }
 
 type aggregate struct {
-	participationManager ParticipationManager
+	crowdactionParticipationManager CrowdactionParticipationManager
 }
 
-func NewParticipationAggregationService(participationManager ParticipationManager) Service {
-	return &aggregate{participationManager: participationManager}
+func NewParticipationAggregationService(crowdactionParticipationManager CrowdactionParticipationManager) Service {
+	return &aggregate{crowdactionParticipationManager: crowdactionParticipationManager}
 }
 
 func (e *aggregate) ChangeCrowdactionParticipantCountBy(ctx context.Context, crowdactionID string, count int) error {
-	return e.participationManager.ChangeCrowdactionParticipantCountBy(ctx, crowdactionID, count)
+	return e.crowdactionParticipationManager.ChangeCrowdactionParticipantCountBy(ctx, crowdactionID, count)
 }
